+++ apps/build-onchain-apps/pages/_app.tsx
@@ -4,8 +4,8 @@ import '../src/styles/global.css';
 import Head from 'next/head';
 import { ThemeProvider } from 'next-themes';
 import { Theme } from '@radix-ui/themes';
-import OnchainProviders from '../src/onchain/OnchainProviders';
-import { MobileMenuProvider } from '../src/components/MobileMenu';
+import OnchainProviders from '../src/providers/OnchainProviders';
+import MobileMenuProvider from '../src/providers/MobileMenuProvider';
 import { initAnalytics } from '../src/utils/analytics';
 import type { AppProps } from 'next/app';
 
@@ -22,15 +22,15 @@ export default function App({ Component, pageProps }: AppProps) {
 
         <link rel="manifest" href="/manifest.json" />
       </Head>
-      <ThemeProvider disableTransitionOnChange attribute="class" value={themeValues}>
-        <Theme accentColor="orange">
-          <OnchainProviders>
+      <OnchainProviders>
+        <ThemeProvider disableTransitionOnChange attribute="class" value={themeValues}>
+          <Theme accentColor="orange">
             <MobileMenuProvider>
               <Component {...pageProps} />
             </MobileMenuProvider>
-          </OnchainProviders>
-        </Theme>
-      </ThemeProvider>
+          </Theme>
+        </ThemeProvider>
+      </OnchainProviders>
     </>
   );
 }
