+++ apps/build-onchain-apps/src/components/header/Header.tsx
@@ -3,10 +3,10 @@ import { Flex, IconButton, Theme, Tooltip } from '@radix-ui/themes';
 import Image from 'next/image';
 import { GitHubLogoIcon, HamburgerMenuIcon } from '@radix-ui/react-icons';
 import NextLink from 'next/link';
-import { classNames } from '../utils/classNames';
-import { AccountConnectButton } from '../onchain/components';
+import { classNames } from '../../utils/classNames';
+import { AccountConnectButton } from '../../onchain/components';
 import logo from '../../public/logo.svg';
-import { ThemeToggle } from './ThemeToggle';
+import { ThemeToggle } from '../ThemeToggle';
 import styles from './Header.module.css';
 import { DefaultNavbar } from './Navbar';
 import { MobileMenu, useMobileMenuContext } from './MobileMenu';
