+++ apps/build-onchain-apps/src/contract/Custom1155.ts
