+++ apps/build-onchain-apps/src/onchainKit/types.ts
@@ -1,5 +1,3 @@
-export type OnchainAccountAddress = `0x${string}`;
-
 export type WagmiFetchBalanceResult = {
   decimals: number;
   formatted: string;
