+++ apps/build-onchain-apps/src/providers/MobileMenuProvider.tsx
@@ -1,7 +1,5 @@
 import { Dispatch, memo, useState, SetStateAction, ReactNode, useEffect } from 'react';
 import { createContext } from '@radix-ui/react-context';
-import { Flex, Portal, Slot, Theme } from '@radix-ui/themes';
-import { RemoveScroll } from 'react-remove-scroll';
 import { useRouter } from 'next/router';
 
 const [MenuProvider, useMenuContext] = createContext<{
@@ -13,9 +11,8 @@ type MobileMenuProps = {
   children?: ReactNode;
 };
 
-export const MobileMenuProvider = memo<MobileMenuProps>(function MobileMenuProvider({ children }) {
+const MobileMenuProvider = memo<MobileMenuProps>(function MobileMenuProvider({ children }) {
   const [open, setOpen] = useState(false);
-
   const router = useRouter();
 
   useEffect(() => {
@@ -24,12 +21,9 @@ export const MobileMenuProvider = memo<MobileMenuProps>(function MobileMenuProvi
       if (document.activeElement instanceof HTMLInputElement) {
         document.activeElement.blur();
       }
-
       setOpen(false);
     };
-
     router.events.on('routeChangeStart', handleRouteChangeStart);
-
     return () => {
       router.events.off('routeChangeStart', handleRouteChangeStart);
     };
@@ -38,11 +32,9 @@ export const MobileMenuProvider = memo<MobileMenuProps>(function MobileMenuProvi
   useEffect(() => {
     // Match @media (--md)
     const mediaQueryList = window.matchMedia('(min-width: 768px)');
-
     const handleChange = () => {
       setOpen((o) => (o ? !mediaQueryList.matches : false));
     };
-
     handleChange();
     mediaQueryList.addEventListener('change', handleChange);
     return () => mediaQueryList.removeEventListener('change', handleChange);
@@ -55,30 +47,6 @@ export const MobileMenuProvider = memo<MobileMenuProps>(function MobileMenuProvi
   );
 });
 
-export const useMobileMenuContext = () => useMenuContext('MobileMenu');
+export { useMenuContext };
 
-export const MobileMenu = memo<MobileMenuProps>(function MobileMenu({ children }) {
-  const mobileMenu = useMobileMenuContext();
-
-  if (!mobileMenu.open) return null;
-
-  return (
-    <Portal>
-      <Theme className="radix-themes-custom-fonts">
-        <RemoveScroll as={Slot} allowPinchZoom enabled>
-          <Flex
-            position="fixed"
-            direction="column"
-            inset="0"
-            style={{
-              marginTop: 'var(--header-height)',
-              backgroundColor: 'var(--color-background)',
-            }}
-          >
-            {children}
-          </Flex>
-        </RemoveScroll>
-      </Theme>
-    </Portal>
-  );
-});
+export default MobileMenuProvider;
